Moved toy carcass pool functionality and accompanying tests to its own module to facilitate testing and development by soil module.
