+++ virtual_rainforest/models/animals/dummy_animal_module.py
@@ -65,13 +65,3 @@ class PalatableSoil:
     """The amount of energy in the soil pool [J]."""
     position: int
     """The grid position of the soil pool."""
-
-
-@dataclass
-class CarcassPool:
-    """This is a class of carcass pools."""
-
-    energy: float
-    """The amount of energy in the carcass pool [J]."""
-    position: int
-    """The grid position of the carcass pool."""
